CopyCommitReferenceAction: better usage of ProgressIndicator
I don't know how `logProviders` work,[1] but I suspect that expression
(logProviders.size * hashes.size) is a misleading overall size of the
task.  If I had to guess, then usually each hash corresponds to a single
logProvider.

As a slighly more logical alternative, provide only the progress on
hashes as a fraction to ProgressIndicator.

[1] Maybe it has to do with Git submodule support or something like
    that.