CopyCommitReferenceAction: tweak progress reporting
Currently, ProgressIndicator is used to just report the number of the
processed hashes.  This doesn't cover the other operations done by
action, like putting them into the OS clipboard via CopyPasteManager.

Make it so that passing to the `consumer` (i.e. CopyPasteManager)
counted the same amount as extracting a single hash.