Gradle: disable generation of attribute "until-build"
By default, if `sinceBuild` and `untilBuild` aren't defined, IntelliJ
Gradle plugin uses the version of the currently used IntelliJ with a
wildcard as the default value.  See code of method
`configurePatchPluginXmlTask` in class `IntelliJPlugin`.

The value for "since-build" attribute is already set to `231` (2023.1)
via `gradle.properties`.[1]  All the APIs used by the plugin are likely
to stay stable enough.  Plus, the plugin itself is tiny with regard to
the amount of code.  I don't want to limit the "until-build" attribute.

So set flag `updateSinceUntilBuild` of the IntelliJ Plugin to `false` to
disable the automatic substitution of `untilBuild`.
https://plugins.jetbrains.com/docs/intellij/tools-gradle-intellij-plugin.html#intellij-extension-updatesinceuntilbuild

[1] See usage of property `pluginSinceBuild` in configuration of task
    `patchPluginXml`.