README: tweak header
Remove the word "plugin" from the top-level header of the README file.

JetBrains themselves call their IDEs in two main waiys, it seems:

- "JetBrains IDEs"
  Example: https://blog.jetbrains.com/blog/2021/12/09/jetbrains-ides-and-dotnet-tools-2021-3-released/

- "IntelliJ-based IDEs"
  Example: https://blog.jetbrains.com/blog/2023/04/12/your-jetbrains-all-products-pack-tools-are-ready-to-be-updated-to-2023-1/

Add a paragraph, that _isn't_ included in `plugin.xml`, to describe the
repository in the most basic way possible, using the second of the
options described above.