README: provide two URLs for alpha channel plugin repositories
The URL for alpha channel of the plugin repository from [1] is too wide
in scope.  It adds _all_ the plugins that are published to the alpha
channel to the search results in IntelliJ IDEA.

Provide a more narrow alternative URL, that results in a repository only
for the Copy Commit Reference plugin.

[1] https://plugins.jetbrains.com/docs/intellij/publishing-plugin.html#specifying-a-release-channel