support IntelliJ IDEA 2020.3 and later
Make the plugin compatible with IntelliJ IDEA 2020.3 and later versions
by using older APIs in class CopyRevisionNumberAction and declaring
attribute `since-build="203"` for `plugin.xml`.

Remove override of method getActionUpdateThread() and corresponding
imports.  This method was added to the platform in version 2022.2.
In version 2022.2 IntelliJ IDEA was also switched from running on JRE 11
to JRE 17.[1]  This affects version of class files, so use jvmToolchain
of version 11 in the Gradle build script.

In the future, if needed, using two development branches should be
possible to support both [2020.3, X) and [X, ∞), where X is the version
that would break compatibility, most probably due to removal of the
interface UpdateInBackground, which has been deprecated for removal
since 223 (aka 2022.3).[2]  In such case the current branch `main` will
probably support the newer version X, while a new separate branch would
be created to support the old versions.  The `main` branch will have to
reintroduce the override of `getActionUpdateThread()`.

[1] https://www.jetbrains.com/idea/whatsnew/2022-2/
    "As of v2022.2, IntelliJ IDEA uses JetBrains Runtime 17"
[2] First, it was just deprecated in commit
      https://github.com/JetBrains/intellij-community/commit/418766c70ebec13b01d89638bb838b1ebbffe127
    then deprecated for removal in commit
      https://github.com/JetBrains/intellij-community/commit/1442370290fbda9a81dbf163381c8ea58e373e53