Gradle: patchPluginXml: link to the full changelog
When I browse for plugins or check plugin updates, I often want to look
at the full changelog.  For example, this is useful when upgrading more
than two versions -- the changelog of the skipped version won't be
visible in the UI.

So add a link to the full changelog in plugin.xml for every release.
Append a section "See also" with a single list item via function `plus`
of class org.jetbrains.changelog.Changelog.Item.