GitHub: Release: tweak CHANGELOG pull requests
Tweak the formatting of the pull requests automatically generated by the
GitHub Actions workflow `release.yml`.  Cut the letter "v" at the front
of the name of the tag, so that generated commit message and PR title
don't contain it.  Spell "CHANGELOG" in all caps, same as the filename.
Drop unnecessary word "update" from commit message and PR title.