GitHub: Build: tweak title of draft releases
Make draft release titles similar to the messages in tags v0.2.0-alpha
and v0.1.0-alpha.1, which are inspired by release tags of git.git.  For
example, tag message of v2.41.0 is `Git 2.41`, with capitalized project
name and without the letter "v" in front of the version string.[1]

Add the name of the plugin to the title of draft releases generated by
GitHub Actions workflow "Build".  Drop the unnecessary letter "v" from
the version string in the release draft title, similarly to previous
commit for the Release workflow.

[1] https://github.com/git/git/releases/tag/v2.41.0