README: link to Marketplace for manual installation
I'm not sure if releases generated by GitHub Actions would contain the
plugin's zip file needed for manual installation.  I did upload it
manually for v0.1.0-alpha.1 [1] and v0.2.0-alpha [2].  However, the
plugin's page [3] on JetBrains' Marketplace website will always have a
download button to the actual plugin.

Link to the plugin's page on JetBrains' Marketplace in the instructions
for manual installation of the plugin in README.md.

Maybe in the future links both for GitHub releases and the Marketplace
could be provided.

[1] https://github.com/rybak/intellij-copy-commit-reference/releases/tag/v0.1.0-alpha.1
[2] https://github.com/rybak/intellij-copy-commit-reference/releases/tag/v0.2.0-alpha
[3] https://plugins.jetbrains.com/plugin/22138-copy-commit-reference