Revert "Gradle: patchPluginXml: show version header"
This reverts commit 1db95b9e52676389d312a92e9644664aa6e44bd2.

Unfortunately, because of the way GitHub Actions are setup in the
template, the "nice header" for the "What's New" tab that I expected
while making commit [1] won't be there -- the header will always be
"Unreleased" instead.  This is because in this release process file
CHANGELOG.md is patched _after_ the release has been created, not during
the release process.

[1] 1db95b9 (Gradle: patchPluginXml: show version header, 2023-07-05)