Gradle: patchPluginXml: provide header manually
In previous commit, Gradle Changelog Plugin configuration was changed to
exclude the version string header from the changelog generated by the
plugin.  As explained in commit 1db95b9 (Gradle: patchPluginXml: show
version header, 2023-07-05), I still want to see the header in the
plugin's "What's New" tab in IntelliJ IDEA's GUI.  This tab comes from
`plugin.xml` file in the plugin's jar.  The contents of this file are
partially generated by Gradle task `patchPluginXml`, which belons to a
different plugin, Gradle IntelliJ Plugin.

I don't want to figure out how the Changelog Plugin works to be able to
adjust the generated changelog in the way that I want it to appear (see
previous commit for why it's complicated).

Add the version string header to the <change-notes> tag in `plugin.xml`
directly in the Gradle task `patchPluginXml`, simply via String
concatenation.