README: link to GitHub and Marketplace for manual installation
As mentioned in commit 5e9e141 (README: link to Marketplace for manual
installation, 2023-07-07), GitHub Actions workflow "Release" indeed
attaches the plugin's zip files (everything from build/distributions
directory) to the release page in the build step "Upload Release Asset"
in `release.yml`.

So provide links to both the latest release on GitHub and to the
versions page on the JetBrains Marketplace in the manual installation
instructions in README.md.