CopyCommitReferenceAction: support issue links with HTML content
Mention of ticket keys, pull request numbers, etc. can be converted into
clickable links in the GUI of IntelliJ IDEA via "Issue Navigation" [1].
This makes it easy to navigate to the issue tracker or VCS server.

Use class IssueLinkHtmlRenderer to produce HTML content of the commit
reference with links according to project's Issue Navigation settings.
Provide both HTML and plain text variants of the commit reference to the
CopyPasteManager.

[1] https://www.jetbrains.com/help/idea/settings-version-control-issue-navigation.html