README: add gif demo
Rendered Markdown files on GitHub on a wide monitor are limited to the
width of 830 pixels in the directory view ("tree") and to 1012 pixels in
the single file view ("blob").

Add a gif demonstrating how the plugin can be used to insert a commit
reference into a commit message.  Show the gif in the README.md file.
Use a gif of resolution 830x750 pixels to fit into both widths.

The original recording was done at 1000x750 using Lossless settings in
OBS Studio.[1]  The raw footage was edited and cropped in Kdenlive.[2]
The edited video was rendered using Lossless FFV1 codec.  The rendered
video file was converted into a gif using an ffmpeg command [3] adapted
from a superuser.com answer [4]:

    ffmpeg -i kdenlive-render.mkv \
        -vf 'split[s0][s1];[s0]palettegen[p];[s1][p]paletteuse' \
        intellij-copy-commit-reference-demo.gif

[1] https://obsproject.com/
[2] https://kdenlive.org/
[3] https://ffmpeg.org/
[4] https://superuser.com/a/556031/133421