CopyCommitReferenceAction: tweak error message text
For the actionName used as background task, which is shown to the user
with a progress meter, we use the `templateText` of the `AnAction`,
which is a super-class of `CopyCommitReferenceAction`.  It is nullable,
and just as a precaution, we use an error message indicating that the
bundle (resource file with key-text pairs) is broken.

Change this error text in anticipation of introducing another action to
the plugin.