CopyCommitReferenceAnnotationAction: new action
Turns out that it is possible to add plugin's action to the context menu
of annotations in IntelliJ editor, and my assumptions about the
construction of that menu were incorrect.

Add CopyCommitReferenceGutterActionProvider to plugin.xml via the
<vcsAnnotationGutterActionProvider> extension.[1]  This class creates
instances of CopyCommitReferenceAnnotationAction.  Aside from the
obvious copying of the commit reference for the given annotation line,
this action also keeps track of the state of the annotations via
interface UpToDateLineNumberListener.[2]

Replace the incorrect assumption in README.md with a link to the TODO.md
file on the `todo` branch of the GitHub repository.

[1] Full list of extensions for plugins is available in JetBrains docs
    https://plugins.jetbrains.com/docs/intellij/plugin-extensions.html
    Extension <vcsAnnotationGutterActionProvider> is used by method
    com.intellij.openapi.vcs.actions.AnnotateToggleAction#addActionsFromExtensions
    I've originally found the extension in source code of plugin
    "Bitbucket Linky"
    https://bitbucket.org/atlassianlabs/intellij-bitbucket-references-plugin/src/0dbcdc83399faf933566ed1e238113f02d0ae45d/src/main/resources/META-INF/plugin.xml?at=master#lines-46:51
[2] The passing of up-to-date line numbers happens in method
    com.intellij.openapi.vcs.actions.AnnotationPresentation#getActions(int)