GitHub: Build: ignore existing ~/.gradle
As part of workflow "Build", step "Disable Gradle Welcome" creates the
directory "~/.gradle".  Apparently, it's possible to have directory
~/.gradle/ to already exist on the runner where workflow is being run:

    mkdir: cannot create directory ‘/home/runner/.gradle’: File exists
    Error: Process completed with exit code 1.

This issue has existed since this step was introduced in commit 8afbba2
(GitHub: Build: disable welcome message from Gradle, 2023-07-06).  Add
flag `-p` to the invocation of `mkdir` to not produce an error if the
directory already exists.